+++ Headers/config_backup.h
@@ -1,0 +1,9 @@
+#ifndef PDW_CONFIG_BACKUP_H
+#define PDW_CONFIG_BACKUP_H
+
+#include <windows.h>
+
+void ShowConfigurationBackupDialog(HWND owner);
+bool ConfigurationRestoreCompleted(void);
+
+#endif